Move manifest data definitions from manifest-updater to key-rotator.
I plan to move functionality out of manifest-updater & into key-rotator
as I start to use it. The key-rotator directory is eventually going
away, and this will help me ensure that I review code as I import it.

Also, bump all Go projects to Go 1.17 (from 1.15).
