+++ .github/workflows/ci-build.yml
@@ -114,7 +114,7 @@ jobs:
     - name: Setup go
       uses: actions/setup-go@v1
       with:
-        go-version: 1.15
+        go-version: 1.17
     - name: Get kubebuilder
       run: |
         wget https://github.com/kubernetes-sigs/kubebuilder/releases/download/v2.3.1/kubebuilder_2.3.1_linux_amd64.tar.gz
@@ -154,7 +154,7 @@ jobs:
     - name: Setup go
       uses: actions/setup-go@v1
       with:
-        go-version: 1.15
+        go-version: 1.17
     - name: Test
       run: go test -race --coverprofile=cover.out --covermode=atomic ./...
     - name: Upload test coverage
@@ -176,7 +176,7 @@ jobs:
     - name: Setup go
       uses: actions/setup-go@v1
       with:
-        go-version: 1.15
+        go-version: 1.17
     - name: Test
       run: go test -race --coverprofile=cover.out --covermode=atomic ./...
     - name: Upload test coverage
