+++ key-rotator/manifest/manifest.go
